fix(tauri) refactor buiilds
