--- tauri/build.rs
@@ -1,5 +1,0 @@
-use tauri_build::build;
-
-fn main() {
-  build();
-}
